requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
cloudinary==1.29.0
dj-database-url==0.5.0
Django==4.0.5
django-cloudinary-storage==0.3.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.1.1
psycopg2==2.9.3
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.2.0
